Добавлена поддержка платформы amd64, реализован режим, при котором два интерфейса TCPMASTER и RTUMASTER работают в одной связке для передачи данных между двумя Modbus Slave устройствами. Работа не закончена, нерешённые вопросы приведены в файле "notes.txt".

+++ clients.c
@@ -249,7 +249,7 @@ int gateway_common_processing()
 	struct timeval tv, tv_wait;
 	struct timezone tz;
 
-	int arg;
+	long arg;
 
 	FD_ZERO(&watchset);
