+++ iface_rtuslave.c
@@ -27,8 +27,8 @@ void *iface_rtu_slave(void *arg)
 
 	int status;
 
-  int port_id=((unsigned)arg)>>8;
-  int client_id=((unsigned)arg)&0xff;
+  int port_id=((long)arg)>>8;
+  int client_id=((long)arg)&0xff;
 
 	GW_Iface	*rtu_slave;
